greentech/requirements.txt
Django==5.0.2
graphene-django==3.1.0
django-graphql-jwt==0.4.0
Pillow==10.2.0
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.34
python-dotenv==1.0.1
django-filter==24.1 